Fix plugin so that it can be loaded multiple times.
Plugin is loaded each time it is installed (automatically), meaning it must be
loaded multiple times within a single execution of the ExpireSubscriptions
utility.

This is accomplished by moving ensuring that functions and constants are not
defined twice, i.e. plugin hook functions are moved to a separate file and
require_once'd, and constant definitions are wrapped in a defined() check.